added pixel art data and created static pixel art component
