+++ src/data/images.json
@@ -1,0 +1,14 @@
+[
+  {
+    "id": "1",
+    "touchedPixels": "[[4, 5], [4, 9], [8, 4], [8, 5], [9, 5], [9, 6], [9, 7], [9, 8], [9, 9], [8, 9], [8, 10], [1, 6], [1, 5], [1, 4], [1, 3], [2, 3], [2, 2], [3, 1], [4, 1], [5, 0], [6, 0], [7, 0], [8, 0], [9, 0], [10, 0], [10, 1], [11, 1], [12, 1], [12, 2], [13, 2], [13, 3], [14, 3], [14, 4], [14, 5], [14, 6], [14, 7], [14, 8], [14, 9], [14, 10], [13, 11], [13, 12], [12, 12], [12, 13], [11, 13], [10, 13], [9, 13], [8, 13], [7, 13], [6, 13], [5, 13], [4, 13], [3, 13], [3, 12], [2, 12], [2, 11], [2, 10], [1, 10], [1, 9], [1, 8], [1, 7]]"
+  },
+  {
+    "id": "2",
+    "touchedPixels": " [[4, 6], [3, 6], [2, 5], [2, 4], [2, 3], [2, 2], [3, 2], [3, 1], [4, 1], [5, 1], [6, 1], [7, 2], [8, 2], [8, 3], [9, 3], [10, 4], [11, 5], [11, 6], [12, 7], [12, 8], [4, 8], [3, 8], [3, 9], [3, 10], [2, 10], [2, 11], [2, 12], [3, 12], [4, 12], [5, 12], [6, 12], [7, 12], [8, 12], [8, 11], [9, 11], [9, 10], [10, 10], [10, 9], [11, 9], [11, 8], [5, 7]]"
+  },
+  {
+    "id": "3",
+    "touchedPixels": " [[4, 4], [4, 9], [9, 4], [9, 5], [9, 6], [9, 8], [9, 9], [9, 10], [9, 11], [9, 7], [9, 3]]"
+  }
+]
