+++ src/views/CreatePixelArtStack/PostPixelArt.tsx
@@ -1,54 +1,6 @@
 import React, {useState} from 'react';
 import {View, Text, StyleSheet} from 'react-native';
-
-function Pixel({width, touched}: {width: number; touched: boolean}) {
-  const [pixelColor, setPixelColor] = useState(touched ? '#fff' : 'black');
-  return (
-    <View
-      style={{
-        width: width,
-        borderWidth: 0.25,
-        borderColor: '#404040',
-        backgroundColor: pixelColor,
-      }}
-    />
-  );
-}
-
-interface CreateGridProps {
-  gridSize: number;
-  gridWidth: number;
-  touchedPixels: [] | [number, number][];
-}
-
-function createGrid({gridSize, gridWidth, touchedPixels}: CreateGridProps) {
-  const rows = [];
-  const pixelWidth = gridWidth / gridSize;
-  for (let i = 0; i < gridSize; i++) {
-    const pixels = [];
-
-    for (let j = 0; j < gridSize; j++) {
-      const touched = touchedPixels.some(
-        elem => elem[0] === i && elem[1] === j,
-      );
-      pixels.push(
-        <Pixel width={pixelWidth} key={`${i}-${j}`} touched={touched} />,
-      );
-    }
-
-    rows.push(
-      <View
-        key={`row-${i}`}
-        style={{
-          flex: 1,
-          flexDirection: 'row',
-        }}>
-        {pixels}
-      </View>,
-    );
-  }
-  return rows;
-}
+import {createGrid} from '../../components/StaticPixelArt';
 
 const gridWidth = 100; //change this to a relative value not abs
